requirements.txt
requests==2.26.0
pillow==10.4.0
moviepy==1.0.3
numpy==2.1.2
pytest==8.3.3